refactored code and broke the product entity apart into separate file
- ran all tests, they all passed
