+++ vend.js
@@ -6,301 +6,8 @@ var Promise = null;
 var request = null;
 var log = null;
 
-/* jshint ignore:start */
-function RateLimitingError(e) {
-  return e.statusCode == 429;
-}
-
-function AuthNError(e) {
-  return e.statusCode == 401;
-}
-
-function AuthZError(e) {
-  return e.statusCode == 403;
-}
-
-function ClientError(e) {
-  return e.statusCode >= 400 && e.statusCode < 500;
-}
-
-function ConnectError(e) { // TODO: maybe leverage https://github.com/petkaantonov/core-error-predicates#connecterror
-  return e.code === 'ETIMEDOUT';
-}
-/* jshint ignore:end */
-
-var successHandler = function(response) {
-  if(_.isArray(response)) {
-    log.debug('response is an array');
-  }
-  else if(_.isObject(response)) {
-    log.debug('response is an object');
-    return Promise.resolve(response);
-  }
-  else if(_.isString(response)) {
-    log.debug('response is a string');
-    try{
-      var responseObject = JSON.parse(response);
-      //log.silly(responseObject);
-      return Promise.resolve(responseObject);
-    }
-    catch(error){
-      log.error('successHandler', 'caught an error: ', error);
-      throw error;
-    }
-  }
-  else {
-    log.debug(response);
-  }
-};
-
-var retryWhenConnectionFails = function(args, connectionInfo, callback, retryCounter) {
-  if(retryCounter<3) {
-    var retryAfter = (retryCounter+1)*1000; // will wait for 1, 2, 3 seconds successively
-    log.debug('retryWhenConnectionFails', 'retry after: ' + retryAfter + ' ms');
-
-    return Promise.delay(retryAfter)
-        .then(function() {
-          log.debug('retryWhenConnectionFails', retryAfter + ' ms have passed...');
-          return callback(args, connectionInfo, ++retryCounter);
-        });
-  }
-  else {
-    return Promise.reject('failed to connect, even after multiple retries');
-  }
-};
-
-/**
- * TODO: Should we reuse the following library instead of rolling our own implementation here?
- *       https://github.com/you21979/node-limit-request-promise
- *
- * @param bodyObject
- * @param connectionInfo - contains tokens and domainPrefix
- * @param retryCounter
- * @param callback
- * @returns {*|Parse.Promise}
- */
-var retryWhenRateLimited = function(bodyObject, args, connectionInfo, callback, retryCounter) {
-  if(retryCounter<3) {
-    var retryAfter = 5*60*1000; // by default Vend will never block for more than 5 minutes
-    retryAfter = Math.max(moment(bodyObject['retry-after']).diff(moment()), 0);
-    //retryAfter = 5000; // for sanity testing counter increments quickly
-    log.debug('retry after: ' + retryAfter + ' ms');
-
-    return Promise.delay(retryAfter)
-      .then(function() {
-        log.debug(retryAfter + ' ms have passed...');
-        return callback(args, connectionInfo, ++retryCounter);
-      });
-  }
-};
-
-var retryWhenAuthNFails = function(args, connectionInfo, callback, retryCounter) {
-  if(retryCounter<3) {
-    if ( !(connectionInfo.vendTokenService &&
-           connectionInfo.vendClientId &&
-           connectionInfo.vendClientSecret &&
-           connectionInfo.refreshToken) )
-    {
-      return Promise.reject('missing required arguments for retryWhenAuthNFails()');
-      // throw e; // TODO: throw unknown errors but reject well known errors?
-    }
-
-    log.debug('try to fetch a new access token');
-    return refreshAccessToken( //TODO: validate connectionInfo before using it for retries?
-      connectionInfo.vendTokenService,
-      connectionInfo.vendClientId,
-      connectionInfo.vendClientSecret,
-      connectionInfo.refreshToken,
-      connectionInfo.domainPrefix
-    )
-      .then(function(oauthInfo) {/*jshint camelcase: false */
-        log.debug('update connectionInfo w/ new token before using it again', oauthInfo);
-        var waitFor = Promise.resolve();
-        if (oauthInfo.access_token) {
-          log.debug('received new access_token: ' + oauthInfo.access_token);
-          connectionInfo.accessToken = oauthInfo.access_token;
-          if(_.isFunction(connectionInfo.updateAccessToken)) {
-            waitFor = connectionInfo.updateAccessToken(connectionInfo);
-          }
-        }
-        if (oauthInfo.refresh_token) {
-          log.debug('received new refresh_token: ' + oauthInfo.refresh_token);
-          connectionInfo.refreshToken = oauthInfo.refresh_token;
-        }
-
-        log.debug('retrying with new accessToken: ' + connectionInfo.accessToken);
-        return waitFor.then(function(){
-          return callback(args, connectionInfo, ++retryCounter);
-        });
-      });
-  }
-};
-
-var sendRequest = function(options, args, connectionInfo, callback, retryCounter) {
-  if ( !(connectionInfo && connectionInfo.accessToken && connectionInfo.domainPrefix) ) {
-    return Promise.reject('missing required arguments for sendRequest()');
-  }
-  if (options.headers) {
-    options.headers['User-Agent'] = process.env['User-Agent'] + '.vend-nodejs-sdk';
-  }
-  return request(options)
-    .then(successHandler)
-    .catch(ConnectError, function(e) {// jshint ignore:line
-      log.error('A ConnectError happened: \n' + e);
-      return retryWhenConnectionFails(args, connectionInfo, callback, retryCounter);
-      // TODO: how to prevent a throw or rejection from also stepping thru the other catch-blocks?
-    })
-    .catch(RateLimitingError, function(e) {// jshint ignore:line
-      log.error('A RateLimitingError error like "429 Too Many Requests" happened: \n'
-          + 'statusCode: ' + e.statusCode + '\n'
-          + 'body: ' + e.response.body + '\n'
-        //+ JSON.stringify(e.response.headers,null,2)
-      );
-
-      var bodyObject = JSON.parse(e.response.body);
-      log.debug(bodyObject['retry-after']);
-      log.debug(
-        moment(bodyObject['retry-after']).format('dddd, MMMM Do YYYY, h:mm:ss a ZZ')
-      );
-      /*successHandler(e.response.body)
-        .then(function(bodyObject){
-          log.debug(bodyObject['retry-after']);
-          log.debug(
-            moment(bodyObject['retry-after']).format('dddd, MMMM Do YYYY, h:mm:ss a ZZ')
-          );
-        });*/
-      return retryWhenRateLimited(bodyObject, args, connectionInfo, callback, retryCounter);
-      // TODO: how should a catch-block respond if there is a problem within the retry?
-    })
-    .catch(AuthNError, function(e) {// jshint ignore:line
-      log.error('An AuthNError happened: \n'
-          + 'statusCode: ' + e.statusCode + '\n'
-          + 'body: ' + e.response.body + '\n'
-        /*+ JSON.stringify(e.response.headers,null,2)
-         + JSON.stringify(e,null,2)*/
-      );
-      return retryWhenAuthNFails(args, connectionInfo, callback, retryCounter);
-      // TODO: how to prevent a throw or rejection from also stepping thru the other catch-blocks?
-    })
-    .catch(ClientError, function(e) {// jshint ignore:line
-      var message = e.response.body;
-      if(_.isObject(message)) {
-        message = JSON.stringify(message,null,2);
-      }
-      log.error('A ClientError happened: \n'
-          + e.statusCode + ' ' + message + '\n'
-        /*+ JSON.stringify(e.response.headers,null,2)
-         + JSON.stringify(e,null,2)*/
-      );
-
-      // TODO: add retry logic
-
-      return Promise.reject(e.statusCode + ' ' + e.response.body); // TODO: throw unknown errors but reject well known errors?
-    })
-    .catch(function(e) {
-      log.error('vend.js - sendRequest - An unexpected error occurred: ', e);
-      throw e; // TODO: throw unknown errors but reject well known errors?
-    });
-};
-
-/**
- * If tokenService already has a domainPrefix set because the API consumer passed in a full URL
- * instead of a substitutable one ... then the replace acts as a no-op.
- *
- * @param tokenService
- * @param domain_prefix
- * @returns {*|XML|string|void}
- */
-var getTokenUrl = function(tokenService, domainPrefix) {
-  var tokenUrl = tokenService.replace(/\{DOMAIN_PREFIX\}/, domainPrefix);
-  log.debug('token Url: '+ tokenUrl);
-  return tokenUrl;
-};
-
-function processPagesRecursively(args, connectionInfo, fetchSinglePage, processPagedResults, previousProcessedResults){
-  return fetchSinglePage(args, connectionInfo)
-    .then(function(result){/*jshint camelcase: false */
-
-      // HACK - until Vend responses become consistent
-      if (result && result.results && !result.pagination) {
-        result.pagination = {
-          'results': result.results,
-          'page': result.page,
-          'page_size': result.page_size,
-          'pages': result.pages,
-        }; // NOTE: if the first page has all the results, this block won't run then either
-      }
-
-      if(result.pagination && result.pagination.pages > args.page.value) {
-        log.info('# of entries returned: ' + result.pagination.results);
-        log.info('Page # ' + args.page.value + ' of ' + result.pagination.pages);
-        return processPagedResults(result, previousProcessedResults)
-          .then(function(newlyProcessedResults){
-            args.page.value = args.page.value+1;
-            return processPagesRecursively(args, connectionInfo, fetchSinglePage, processPagedResults, newlyProcessedResults);
-          });
-      }
-      else if (result && result.version && result.data && result.data.length>0) {
-        log.info('# of entries returned: ' + result.data.length);
-        log.info('version min: '+ result.version.min + ' max: ' + result.version.max);
-        if (args.pageSize.value) {
-          log.info('Page # ' + args.page.value); // page has no operational role here, just useful for readable logs
-        }
-        return processPagedResults(result, previousProcessedResults)
-          .then(function(newlyProcessedResults){
-            args.after.value = result.version.max; // work on whatever is left after the max version from previous call
-            args.page.value = args.page.value+1; // page has no operational role here, just useful for readable logs
-            return processPagesRecursively(args, connectionInfo, fetchSinglePage, processPagedResults, newlyProcessedResults);
-          });
-      }
-      else {
-        log.info('Processing last page. Page # ' + args.page.value);
-        return processPagedResults(result, previousProcessedResults);
-      }
-    });
-}
-
-var processPromisesSerially = function(aArray, aArrayIndex, args, mergeStrategy, setupNext, executeNext, aPreviousResults){
-  if (aArrayIndex < aArray.length) {
-    log.debug('processPromisesSerially for aArrayIndex # ' + aArrayIndex);
-    return executeNext(args)
-      .then(function(executedResults){
-        //log.silly('executedResults ', executedResults);
-        //log.silly('executedResults.length ', executedResults.length); // .length may not be valid everytime
-        return mergeStrategy(executedResults, aPreviousResults, args)
-          .then(function(mergedResults){
-            log.debug('mergedResults.length ', mergedResults.length);
-            //log.silly('before: ', args);
-            args = setupNext(args);
-            //log.silly('after: ', args);
-            return processPromisesSerially(
-              args.getArray(), //args.consignmentIds.value,
-              args.getArrayIndex(), //args.consignmentIdIndex.value,
-              args,
-              mergeStrategy,
-              setupNext,
-              executeNext,
-              mergedResults
-            );
-          });
-      });
-  }
-  else {
-    if(aPreviousResults) {
-    log.debug('aPreviousResults.length ', aPreviousResults.length);
-    }
-    log.debug('processPromisesSerially() finished');
-    return Promise.resolve(aPreviousResults);
-  }
-};
-
-var argsAreValid = function(args){
-  var arrayOfRequiredArgs = _.filter(args, function(object/*, key*/){
-    return object.required;
-  });
-  var arrayOfRequiredValues = _.pluck(arrayOfRequiredArgs, 'value');
-  return !_.contains(arrayOfRequiredValues, undefined);
-};
+var utils = null;
+var product = null;
 
 // the API consumer will get the args and fill in the blanks
 // the SDK will pull out the non-empty values and execute the request
@@ -462,95 +169,6 @@ var argsForInput = {
           }
         };
       }
-    }
-  },
-  products: {
-    fetchById: function() {
-      return {
-        apiId: {
-          required: true,
-          value: undefined
-        }
-      };
-    },
-    update: function() {
-      return {
-        body: {
-          required: true,
-          value: undefined
-        }
-      };
-    },
-    create: function() {
-      return {
-        body: {
-          required: true,
-          value: undefined
-        }
-      };
-    },
-    uploadImage: function() {
-      return {
-        apiId: {
-          required: true,
-          value: undefined
-        },
-        image: {
-          required: true,
-          value: undefined
-        }
-      };
-    },
-    fetch: function() {
-      return {
-        orderBy: {
-          required: false,
-          key: 'order_by',
-          value: undefined // updated_at (default) | id | name
-        },
-        orderDirection: {
-          required: false,
-          key: 'order_direction',
-          value: undefined // ASC (default) | DESC
-          //TODO: setup enumerations in javascript?
-        },
-        since: {
-          required: false,
-          key: 'since',
-          value: undefined
-        },
-        active: {
-          required: false,
-          key: 'active',
-          value: undefined // 0 (or no value) : returns only inactive products
-                           // 1 (or any other value) : returns only active products
-          // TODO: can we embed a transformation here?
-          //       API consumer will set true or false or 0 or 1 as the value
-          //       but SDK will get the 0 or 1 value based on a transformation
-        },
-        page: {
-          required: false,
-          key: 'page',
-          value: undefined
-        },
-        pageSize: {
-          required: false,
-          key: 'page_size',
-          value: undefined
-        }
-      };
-    },
-    fetchAll: function() {
-      return {
-        active: {
-          required: false,
-          key: 'active',
-          // 0 (or no value) : returns only inactive products
-          // 1 (or any other value) : returns only active products
-          // undefined : returns both active and inactive products
-          value: undefined
-        },
-      };
     }
   },
   customers: {
@@ -908,7 +526,7 @@ var fetchStockOrdersForSuppliers = function(args, connectionInfo, retryCounter) 
     }
   };
 
-  return sendRequest(options, args, connectionInfo, fetchStockOrdersForSuppliers, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, fetchStockOrdersForSuppliers, retryCounter);
 };
 
 var fetchAllStockOrdersForSuppliers = function(connectionInfo, processPagedResults) {
@@ -932,7 +550,7 @@ var fetchAllStockOrdersForSuppliers = function(connectionInfo, processPagedResul
       return Promise.resolve(pagedData.consignments);
     };
   }
-  return processPagesRecursively(args, connectionInfo, fetchStockOrdersForSuppliers, processPagedResults);
+  return utils.processPagesRecursively(args, connectionInfo, fetchStockOrdersForSuppliers, processPagedResults);
 };
 
 var fetchProductsByConsignment  = function(args, connectionInfo, retryCounter) {
@@ -962,7 +580,7 @@ var fetchProductsByConsignment  = function(args, connectionInfo, retryCounter) {
     }
   };
 
-  return sendRequest(options, args, connectionInfo, fetchProductsByConsignment, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, fetchProductsByConsignment, retryCounter);
 };
 
 var defaultMethod_ForProcessingPagedResults_ForConsignmentProducts = function(pagedData, previousData){// jshint ignore:line
@@ -1007,7 +625,7 @@ var fetchAllProductsByConsignment = function(args, connectionInfo, processPagedR
   if (!processPagedResults) {
     processPagedResults = defaultMethod_ForProcessingPagedResults_ForConsignmentProducts;// jshint ignore:line
   }
-  return processPagesRecursively(args, connectionInfo, fetchProductsByConsignment, processPagedResults);
+  return utils.processPagesRecursively(args, connectionInfo, fetchProductsByConsignment, processPagedResults);
 };
 
 var fetchAllProductsByConsignments = function(args, connectionInfo, processPagedResults) {
@@ -1024,7 +642,7 @@ var fetchAllProductsByConsignments = function(args, connectionInfo, processPaged
   };
 
   // iterate serially through a promise chain for all args.consignmentIds.value
-  return processPromisesSerially(
+  return utils.processPromisesSerially(
     args.consignmentIds.value,
     args.consignmentIdIndex.value,
     args,
@@ -1079,7 +697,7 @@ var resolveMissingSuppliers = function(args, connectionInfo) {
   };
 
   // iterate serially through a promise chain for all args.consignmentIds.value
-  return processPromisesSerially(
+  return utils.processPromisesSerially(
     args.getArray(),
     args.getArrayIndex(),
     args,
@@ -1114,379 +732,9 @@ var resolveMissingSuppliers = function(args, connectionInfo) {
       //log.silly('updatedArgs: ', updatedArgs);
       var args = argsForInput.products.fetchById();
       args.apiId.value = updatedArgs.consignmentProductId.value;
-      return fetchProduct(args, connectionInfo);
+      return product.fetchProduct(args, connectionInfo);
     }
   );
-};
-
-// WARN: if the ID is incorrect, the vend api the first 50 products which can totally throw folks off their mark!
-// TODO: instead of returning response, return the value of response.products[0] directly?
-var fetchProduct = function(args, connectionInfo, retryCounter) {
-  if (!retryCounter) {
-    retryCounter = 0;
-  } else {
-    log.debug('retry # ' + retryCounter);
-  }
-
-  var path = '/api/products/' + args.apiId.value;
-  // this is an undocumented implementation by Vend
-  // the response has to be accessed like: result.products[0]
-  // which is lame ... TODO: should we unwrap it within the SDK?
-
-  var vendUrl = 'https://' + connectionInfo.domainPrefix + '.vendhq.com' + path;
-  log.debug('Requesting vend product ' + vendUrl);
-  var authString = 'Bearer ' + connectionInfo.accessToken;
-  log.debug('GET ' + vendUrl);
-  log.silly('Authorization: ' + authString); // TODO: sensitive data ... do not log?
-
-  var options = {
-    url: vendUrl,
-    headers: {
-      'Authorization': authString,
-      'Accept': 'application/json'
-    }
-  };
-
-  return sendRequest(options, args, connectionInfo, fetchProduct, retryCounter);
-};
-
-/**
- * This method updates a product by product Id.
- * The product's id is passed in the `body` as a json object along with other parameters
- * instead of passing it in the url as querystring. That's how an update happens in Vend.
- */
-var updateProductById = function(args, connectionInfo, retryCounter) {
-  if ( !(args && argsAreValid(args)) ) {
-    return Promise.reject('missing required arguments for updateProductById()');
-  }
-
-  if (!retryCounter) {
-    retryCounter = 0;
-  } else {
-    log.debug('retry # ' + retryCounter);
-  }
-
-  var path = '/api/products';
-  var vendUrl = 'https://' + connectionInfo.domainPrefix + '.vendhq.com' + path;
-  var authString = 'Bearer ' + connectionInfo.accessToken;
-  log.silly('Authorization: ' + authString); // TODO: sensitive data ... do not log?
-  var body = args.body.value;
-
-  var options = {
-    method: 'POST',
-    url: vendUrl,
-    headers: {
-      'Authorization': authString,
-      'Content-Type': 'application/json',
-      'Accept': 'application/json'
-    },
-    json: body
-  };
-  log.debug(options.method + ' ' + options.url);
-
-  return sendRequest(options, args, connectionInfo, updateProductById, retryCounter);
-};
-
-
-var deleteProductById = function(args, connectionInfo, retryCounter) {
-  log.debug('inside deleteProductById()');
-
-  log.debug(args);
-  if ( !(args && argsAreValid(args)) ) {
-    return Promise.reject('missing required arguments for deleteProductById()');
-  }
-
-  if (!retryCounter) {
-    retryCounter = 0;
-  } else {
-    log.debug('retry # ' + retryCounter);
-  }
-
-  log.debug('args.apiId.value: ' + args.apiId.value);
-  var path = '/api/products/' + args.apiId.value;
-  var vendUrl = 'https://' + connectionInfo.domainPrefix + '.vendhq.com' + path;
-  var authString = 'Bearer ' + connectionInfo.accessToken;
-  log.silly('Authorization: ' + authString); // TODO: sensitive data ... do not log?
-
-  var options = {
-    method: 'DELETE',
-    url: vendUrl,
-    headers: {
-      'Authorization': authString,
-      'Content-Type': 'application/json',
-      'Accept': 'application/json'
-    }
-  };
-  log.debug(options.method + ' ' + options.url);
-
-  return sendRequest(options, args, connectionInfo, deleteProductById, retryCounter);
-};
-
-var createProduct = function(args, connectionInfo, retryCounter) {
-  if ( !(args && argsAreValid(args)) ) {
-    return Promise.reject('missing required arguments for createProduct()');
-  }
-
-  if (!retryCounter) {
-    retryCounter = 0;
-  } else {
-    log.debug('retry # ' + retryCounter);
-  }
-
-  var path = '/api/products';
-  var vendUrl = 'https://' + connectionInfo.domainPrefix + '.vendhq.com' + path;
-  var authString = 'Bearer ' + connectionInfo.accessToken;
-  log.silly('Authorization: ' + authString); // TODO: sensitive data ... do not log?
-  var body = args.body.value;
-
-  var options = {
-    method: 'POST',
-    url: vendUrl,
-    headers: {
-      'Authorization': authString,
-      'Content-Type': 'application/json',
-      'Accept': 'application/json'
-    },
-    json: body
-  };
-  log.debug(options.method + ' ' + options.url);
-
-  return sendRequest(options, args, connectionInfo, createProduct, retryCounter);
-};
-
-var uploadProductImage = function(args, connectionInfo, retryCounter) {
-  if ( !(args && argsAreValid(args)) ) {
-    return Promise.reject('missing required arguments for uploadProductImage()');
-  }
-
-  if (!retryCounter) {
-    retryCounter = 0;
-  } else {
-    log.debug('retry # ' + retryCounter);
-  }
-
-  var path = '/api/2.0/products/' + args.apiId.value + '/actions/image_upload';
-  var vendUrl = 'https://' + connectionInfo.domainPrefix + '.vendhq.com' + path;
-  var authString = 'Bearer ' + connectionInfo.accessToken;
-  log.silly('Authorization: ' + authString); // TODO: sensitive data ... do not log?
-  var body = args.image.value;
-
-  var options = {
-    method: 'POST',
-    url: vendUrl,
-    headers: {
-      'Authorization': authString,
-      'Content-Type': 'application/json',
-      'Accept': 'application/json'
-    },
-    json: body
-  };
-  log.debug(options.method + ' ' + options.url);
-
-  return sendRequest(options, args, connectionInfo, uploadProductImage, retryCounter);
-};
-
-// TODO: instead of returning response, return the value of response.products[0] directly?
-var fetchProductByHandle  = function(args, connectionInfo, retryCounter) {
-  if ( !(args && args.handle && args.handle.value) ) {
-    return Promise.reject('missing required arguments for fetchProductByHandle()');
-  }
-
-  if (!retryCounter) {
-    retryCounter = 0;
-  } else {
-    log.debug('retry # ' + retryCounter);
-  }
-
-  var path = '/api/products';
-  // this is an undocumented implementation by Vend
-  // the response has to be accessed like: result.products[0]
-  // which is lame ... TODO: should we unwrap it within the SDK?
-
-  var vendUrl = 'https://' + connectionInfo.domainPrefix + '.vendhq.com' + path;
-  log.debug('Requesting vend product ' + vendUrl);
-  var authString = 'Bearer ' + connectionInfo.accessToken;
-  log.debug('GET ' + vendUrl);
-  log.silly('Authorization: ' + authString); // TODO: sensitive data ... do not log?
-
-  var options = {
-    url: vendUrl,
-    headers: {
-      'Authorization': authString,
-      'Accept': 'application/json'
-    },
-    qs: {
-      handle: args.handle.value
-    }
-  };
-
-  return sendRequest(options, args, connectionInfo, fetchProductByHandle, retryCounter);
-};
-
-// TODO: instead of returning response, return the value of response.products[0] directly?
-var fetchProductBySku  = function(args, connectionInfo, retryCounter) {
-  if ( !(args && args.sku && args.sku.value) ) {
-    return Promise.reject('missing required arguments for fetchProductByHandle()');
-  }
-
-  if (!retryCounter) {
-    retryCounter = 0;
-  } else {
-    log.debug('retry # ' + retryCounter);
-  }
-
-  var path = '/api/products';
-  // this is an undocumented implementation by Vend
-  // the response has to be accessed like: result.products[0]
-  // which is lame ... TODO: should we unwrap it within the SDK?
-
-  var vendUrl = 'https://' + connectionInfo.domainPrefix + '.vendhq.com' + path;
-  log.debug('Requesting vend product ' + vendUrl);
-  var authString = 'Bearer ' + connectionInfo.accessToken;
-  log.debug('GET ' + vendUrl);
-  log.silly('Authorization: ' + authString); // TODO: sensitive data ... do not log?
-
-  var options = {
-    url: vendUrl,
-    headers: {
-      'Authorization': authString,
-      'Accept': 'application/json'
-    },
-    qs: {
-      sku: args.sku.value
-    }
-  };
-
-  return sendRequest(options, args, connectionInfo, fetchProductBySku, retryCounter);
-};
-
-var fetchProducts = function(args, connectionInfo, retryCounter) {
-  if (!retryCounter) {
-    retryCounter = 0;
-  } else {
-    log.debug('retry # ' + retryCounter);
-  }
-
-  var path = '/api/products';
-  var vendUrl = 'https://' + connectionInfo.domainPrefix + '.vendhq.com' + path;
-  var authString = 'Bearer ' + connectionInfo.accessToken;
-  log.debug('GET ' + vendUrl);
-  log.silly('Authorization: ' + authString); // TODO: sensitive data ... do not log?
-
-  //var domainPrefix = this.domainPrefix;
-
-  var active;
-  if (args.active.value===undefined || args.active.value===null) {
-    active = undefined;
-  } else {
-    active = (args.active.value) ? 1 : 0;
-  }
-  var options = {
-    url: vendUrl,
-    headers: {
-      'Authorization': authString,
-      'Accept': 'application/json'
-    },
-    qs: {/*jshint camelcase: false */
-      order_by: args.orderBy.value,
-      order_direction: args.orderDirection.value,
-      since: args.since.value,
-      active: active,
-      page: args.page.value,
-      page_size: args.pageSize.value
-    }
-  };
-  if (args.page.value) {
-    log.debug('Requesting product page ' + args.page.value);
-  }
-
-  return sendRequest(options, args, connectionInfo, fetchProducts, retryCounter);
-};
-
-var fetchAllProducts = function(args, connectionInfo, processPagedResults) {
-  var defaultArgs = argsForInput.products.fetch();
-  defaultArgs.orderBy.value = 'id';
-  defaultArgs.page.value = 1;
-  defaultArgs.pageSize.value = 200;
-  defaultArgs.active.value = true; // fetch only active products by default
-
-  /*
-   * This method's signature changed
-   * FROM:
-   *     function(connectionInfo, processPagedResults)
-   * TO:
-   *     function(args, connectionInfo, processPagedResults)
-   * therefore, the following section massages arguments to provide backward compatibility.
-   */ 
-  var input = (arguments.length === 1 ? [arguments[0]] : Array.apply(null, arguments));
-  if (input.length===1) { // backward compatible with old method signature
-    connectionInfo = input[0];
-  }
-  else if (input.length===2 && _.isFunction(input[1])) { // backward compatible with old method signature
-    connectionInfo = input[0];
-    processPagedResults = input[1];
-  }
-  else if (
-    ( input.length===2 ) ||
-    ( input.length===3 && _.isFunction(input[2]) ) ) // new method signature calls, will end up using this code-block
-  {
-    if ( !(args && argsAreValid(args)) ) {
-      return Promise.reject('missing required arguments for fetchAllProducts()');
-    }
-    if (!args.active || args.active.value===undefined || args.active.value===null) {
-      // in Vend API, not specifying this field is a way of saying
-      // that you want both: active AND inactive products
-      defaultArgs.active.value = undefined;
-    }
-    else {
-      defaultArgs.active.value = args.active.value;
-    }
-  }
-  else {
-    return Promise.reject('please check the method signature and fix your code');
-  }
-  // method signature related changes END
-
-  // set a default function if none is provided
-  if (!processPagedResults) {
-    processPagedResults = function processPagedResults(pagedData, previousData){
-      log.debug('fetchAllProducts - default processPagedResults()');
-      if (previousData && previousData.length>0) {
-        //log.verbose(JSON.stringify(pagedData.products,replacer,2));
-        if (pagedData.products && pagedData.products.length>0) {
-          log.debug('previousData: ', previousData.length);
-          pagedData.products = pagedData.products.concat(previousData);
-          log.debug('combined: ', pagedData.products.length);
-        }
-        else {
-          pagedData.products = previousData;
-        }
-      }
-      return Promise.resolve(pagedData.products);
-    };
-  }
-  return processPagesRecursively(defaultArgs, connectionInfo, fetchProducts, processPagedResults);
-};
-
-var fetchPaginationInfo = function(args, connectionInfo){
-  if ( !(args && argsAreValid(args)) ) {
-    return Promise.reject('missing required arguments for fetchPaginationInfo()');
-  }
-  return fetchProducts(args, connectionInfo)
-    .then(function(result){/*jshint camelcase: false */
-
-      // HACK - until Vend responses become consistent
-      if (result && result.results && !result.pagination) {
-        result.pagination = {
-          'results': result.results,
-          'page': result.page,
-          'page_size': result.page_size,
-          'pages': result.pages,
-        }; // NOTE: if the first page has all the results, this block won't run then either
-      }
-
-      return (result && result.pagination) ? Promise.resolve(result.pagination) : Promise.resolve();
-    });
 };
 
 var fetchCustomers = function(args, connectionInfo, retryCounter) {
@@ -1518,7 +766,7 @@ var fetchCustomers = function(args, connectionInfo, retryCounter) {
     } //TODO: add page & page_size?
   };
 
-  return sendRequest(options, args, connectionInfo, fetchCustomers, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, fetchCustomers, retryCounter);
 };
 
 var fetchCustomers2 = function(args, connectionInfo, retryCounter) {
@@ -1549,11 +797,11 @@ var fetchCustomers2 = function(args, connectionInfo, retryCounter) {
     }
   };
 
-  return sendRequest(options, args, connectionInfo, fetchCustomers2, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, fetchCustomers2, retryCounter);
 };
 
 var fetchAllCustomers = function(args, connectionInfo, processPagedResults) {
-  if ( !(args && argsAreValid(args)) ) {
+  if ( !(args && utils.argsAreValid(args)) ) {
     return Promise.reject('missing required arguments for createProductTypes()');
   }
   if (!args.page || !args.page.value) {
@@ -1581,7 +829,7 @@ var fetchAllCustomers = function(args, connectionInfo, processPagedResults) {
       // return Promise.resolve();
     };
   }
-  return processPagesRecursively(args, connectionInfo, fetchCustomers2, processPagedResults);
+  return utils.processPagesRecursively(args, connectionInfo, fetchCustomers2, processPagedResults);
 };
 
 var fetchCustomerByEmail = function(email, connectionInfo, retryCounter) {
@@ -1618,7 +866,7 @@ var fetchRegisters = function(args, connectionInfo, retryCounter) {
     }
   };
 
-  return sendRequest(options, args, connectionInfo, fetchRegisters, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, fetchRegisters, retryCounter);
 };
 
 var fetchAllRegisters = function(args, connectionInfo, processPagedResults) {
@@ -1646,7 +894,7 @@ var fetchAllRegisters = function(args, connectionInfo, processPagedResults) {
       return Promise.resolve(pagedData.registers);
     };
   }
-  return processPagesRecursively(args, connectionInfo, fetchRegisters, processPagedResults);
+  return utils.processPagesRecursively(args, connectionInfo, fetchRegisters, processPagedResults);
 };
 
 var fetchRegister  = function(args, connectionInfo, retryCounter) {
@@ -1671,7 +919,7 @@ var fetchRegister  = function(args, connectionInfo, retryCounter) {
     }
   };
 
-  return sendRequest(options, args, connectionInfo, fetchRegister, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, fetchRegister, retryCounter);
 };
 
 var fetchPaymentTypes = function(args, connectionInfo, retryCounter) {
@@ -1697,7 +945,7 @@ var fetchPaymentTypes = function(args, connectionInfo, retryCounter) {
     }
   };
 
-  return sendRequest(options, args, connectionInfo, fetchPaymentTypes, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, fetchPaymentTypes, retryCounter);
 };
 
 var fetchProductTypes = function(args, connectionInfo, retryCounter) {
@@ -1730,11 +978,11 @@ var fetchProductTypes = function(args, connectionInfo, retryCounter) {
     }
   };
 
-  return sendRequest(options, args, connectionInfo, fetchProductTypes, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, fetchProductTypes, retryCounter);
 };
 
 var createProductTypes = function(args, connectionInfo, retryCounter) {
-  if ( !(args && argsAreValid(args)) ) {
+  if ( !(args && utils.argsAreValid(args)) ) {
     return Promise.reject('missing required arguments for createProductTypes()');
   }
 
@@ -1762,7 +1010,7 @@ var createProductTypes = function(args, connectionInfo, retryCounter) {
   };
   log.debug(options.method + ' ' + options.url);
 
-  return sendRequest(options, args, connectionInfo, createProductTypes, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, createProductTypes, retryCounter);
 };
 
 var fetchTaxes = function(args, connectionInfo, retryCounter) {
@@ -1788,11 +1036,11 @@ var fetchTaxes = function(args, connectionInfo, retryCounter) {
     }
   };
 
-  return sendRequest(options, args, connectionInfo, fetchTaxes, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, fetchTaxes, retryCounter);
 };
 
 var createTax = function(args, connectionInfo, retryCounter) {
-  if ( !(args && argsAreValid(args)) ) {
+  if ( !(args && utils.argsAreValid(args)) ) {
     return Promise.reject('missing required arguments for createTax()');
   }
 
@@ -1820,7 +1068,7 @@ var createTax = function(args, connectionInfo, retryCounter) {
   };
   log.debug(options.method + ' ' + options.url);
 
-  return sendRequest(options, args, connectionInfo, createTax, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, createTax, retryCounter);
 };
 
 var fetchBrands = function(args, connectionInfo, retryCounter) {
@@ -1853,11 +1101,11 @@ var fetchBrands = function(args, connectionInfo, retryCounter) {
     }
   };
 
-  return sendRequest(options, args, connectionInfo, fetchBrands, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, fetchBrands, retryCounter);
 };
 
 var createBrand = function(args, connectionInfo, retryCounter) {
-  if ( !(args && argsAreValid(args)) ) {
+  if ( !(args && utils.argsAreValid(args)) ) {
     return Promise.reject('missing required arguments for createBrand()');
   }
 
@@ -1885,7 +1133,7 @@ var createBrand = function(args, connectionInfo, retryCounter) {
   };
   log.debug(options.method + ' ' + options.url);
 
-  return sendRequest(options, args, connectionInfo, createBrand, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, createBrand, retryCounter);
 };
 
 var fetchTags = function(args, connectionInfo, retryCounter) {
@@ -1918,7 +1166,7 @@ var fetchTags = function(args, connectionInfo, retryCounter) {
     }
   };
 
-  return sendRequest(options, args, connectionInfo, fetchTags, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, fetchTags, retryCounter);
 };
 
 var fetchAllTags = function(args, connectionInfo, processPagedResults) {
@@ -1954,11 +1202,11 @@ var fetchAllTags = function(args, connectionInfo, processPagedResults) {
       return Promise.resolve(pagedData.data);
     };
   }
-  return processPagesRecursively(args, connectionInfo, fetchTags, processPagedResults);
+  return utils.processPagesRecursively(args, connectionInfo, fetchTags, processPagedResults);
 };
 
 var createTag = function(args, connectionInfo, retryCounter) {
-  if ( !(args && argsAreValid(args)) ) {
+  if ( !(args && utils.argsAreValid(args)) ) {
     return Promise.reject('missing required arguments for createTag()');
   }
 
@@ -1986,7 +1234,7 @@ var createTag = function(args, connectionInfo, retryCounter) {
   };
   log.debug(options.method + ' ' + options.url);
 
-  return sendRequest(options, args, connectionInfo, createTag, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, createTag, retryCounter);
 };
 
 var fetchRegisterSales = function(args, connectionInfo, retryCounter) {
@@ -2020,7 +1268,7 @@ var fetchRegisterSales = function(args, connectionInfo, retryCounter) {
     }
   };
 
-  return sendRequest(options, args, connectionInfo, fetchRegisterSales, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, fetchRegisterSales, retryCounter);
 };
 
 var fetchAllRegisterSales = function(args, connectionInfo, processPagedResults) {
@@ -2048,7 +1296,7 @@ var fetchAllRegisterSales = function(args, connectionInfo, processPagedResults) 
       return Promise.resolve(pagedData.register_sales);
     };
   }
-  return processPagesRecursively(args, connectionInfo, fetchRegisterSales, processPagedResults);
+  return utils.processPagesRecursively(args, connectionInfo, fetchRegisterSales, processPagedResults);
 };
 
 var fetchOutlets = function(args, connectionInfo, retryCounter) {
@@ -2084,7 +1332,7 @@ var fetchOutlets = function(args, connectionInfo, retryCounter) {
     }
   };
 
-  return sendRequest(options, args, connectionInfo, fetchOutlets, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, fetchOutlets, retryCounter);
 };
 
 var fetchAllOutlets = function(args, connectionInfo, processPagedResults) {
@@ -2121,7 +1369,7 @@ var fetchAllOutlets = function(args, connectionInfo, processPagedResults) {
       return Promise.resolve(pagedData.data);
     };
   }
-  return processPagesRecursively(args, connectionInfo, fetchOutlets, processPagedResults);
+  return utils.processPagesRecursively(args, connectionInfo, fetchOutlets, processPagedResults);
 };
 
 var fetchOutlet = function(args, connectionInfo, retryCounter) {
@@ -2146,7 +1394,7 @@ var fetchOutlet = function(args, connectionInfo, retryCounter) {
     }
   };
 
-  return sendRequest(options, args, connectionInfo, fetchOutlet, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, fetchOutlet, retryCounter);
 };
 
 var fetchSupplier = function(args, connectionInfo, retryCounter) {
@@ -2172,7 +1420,7 @@ var fetchSupplier = function(args, connectionInfo, retryCounter) {
     }
   };
 
-  return sendRequest(options, args, connectionInfo, fetchSupplier, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, fetchSupplier, retryCounter);
 };
 
 var fetchSuppliers = function(args, connectionInfo, retryCounter) {
@@ -2208,7 +1456,7 @@ var fetchSuppliers = function(args, connectionInfo, retryCounter) {
     //       instead of being nested one-level-down under the response.pagination structure!
   }
 
-  return sendRequest(options, args, connectionInfo, fetchSuppliers, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, fetchSuppliers, retryCounter);
 };
 
 var fetchAllSuppliers = function(connectionInfo, processPagedResults) {
@@ -2220,11 +1468,11 @@ var fetchAllSuppliers = function(connectionInfo, processPagedResults) {
   if (!processPagedResults) {
     processPagedResults = defaultMethod_ForProcessingPagedResults_ForSuppliers;// jshint ignore:line
   }
-  return processPagesRecursively(args, connectionInfo, fetchSuppliers, processPagedResults);
+  return utils.processPagesRecursively(args, connectionInfo, fetchSuppliers, processPagedResults);
 };
 
 var createSupplier = function(args, connectionInfo, retryCounter) {
-  if ( !(args && argsAreValid(args)) ) {
+  if ( !(args && utils.argsAreValid(args)) ) {
     return Promise.reject('missing required arguments for createSupplier()');
   }
 
@@ -2252,7 +1500,7 @@ var createSupplier = function(args, connectionInfo, retryCounter) {
   };
   log.debug(options.method + ' ' + options.url);
 
-  return sendRequest(options, args, connectionInfo, createSupplier, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, createSupplier, retryCounter);
 };
 
 var fetchConsignment  = function(args, connectionInfo, retryCounter) {
@@ -2277,7 +1525,7 @@ var fetchConsignment  = function(args, connectionInfo, retryCounter) {
     }
   };
 
-  return sendRequest(options, args, connectionInfo, fetchConsignment, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, fetchConsignment, retryCounter);
 };
 
 var fetchConsignmentProductById  = function(args, connectionInfo, retryCounter) {
@@ -2302,7 +1550,7 @@ var fetchConsignmentProductById  = function(args, connectionInfo, retryCounter) 
     }
   };
 
-  return sendRequest(options, args, connectionInfo, fetchConsignment, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, fetchConsignment, retryCounter);
 };
 
 var createConsignmentProduct = function(args, connectionInfo, retryCounter) {
@@ -2313,7 +1561,7 @@ var createConsignmentProduct = function(args, connectionInfo, retryCounter) {
     body = args.body;
   }
   else {
-    if ( !(args && argsAreValid(args)) ) {
+    if ( !(args && utils.argsAreValid(args)) ) {
       return Promise.reject('missing required arguments for createConsignmentProduct()');
     }
     body = {
@@ -2350,13 +1598,13 @@ var createConsignmentProduct = function(args, connectionInfo, retryCounter) {
   };
   log.debug(options.method + ' ' + options.url);
 
-  return sendRequest(options, args, connectionInfo, createConsignmentProduct, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, createConsignmentProduct, retryCounter);
 };
 
 var createStockOrder = function(args, connectionInfo, retryCounter) {
   log.debug('inside createStockOrder()');
 
-  if ( !(args && argsAreValid(args)) ) {
+  if ( !(args && utils.argsAreValid(args)) ) {
     return Promise.reject('missing required arguments for createStockOrder()');
   }
 
@@ -2392,7 +1640,7 @@ var createStockOrder = function(args, connectionInfo, retryCounter) {
   };
   log.debug(options.method + ' ' + options.url);
 
-  return sendRequest(options, args, connectionInfo, createStockOrder, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, createStockOrder, retryCounter);
 };
 
 var createCustomer = function(body, connectionInfo, retryCounter) {
@@ -2420,11 +1668,11 @@ var createCustomer = function(body, connectionInfo, retryCounter) {
   };
   log.debug(options.method + ' ' + options.url);
 
-  return sendRequest(options, body, connectionInfo, createCustomer, retryCounter);
+  return utils.sendRequest(options, body, connectionInfo, createCustomer, retryCounter);
 };
 
 var fetchRegisterSalesById  = function(args, connectionInfo, retryCounter) {
-  if ( !(args && argsAreValid(args)) ) {
+  if ( !(args && utils.argsAreValid(args)) ) {
     return Promise.reject('missing required arguments for fetchRegisterSalesById()');
   }
 
@@ -2449,7 +1697,7 @@ var fetchRegisterSalesById  = function(args, connectionInfo, retryCounter) {
     }
   };
 
-  return sendRequest(options, args, connectionInfo, fetchRegisterSalesById, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, fetchRegisterSalesById, retryCounter);
 };
 
 var createRegisterSale = function(body, connectionInfo, retryCounter) {
@@ -2484,13 +1732,13 @@ var createRegisterSale = function(body, connectionInfo, retryCounter) {
     json: body
   };
   log.debug(options.method + ' ' + options.url);
-  return sendRequest(options, body, connectionInfo, createRegisterSale, retryCounter);
+  return utils.sendRequest(options, body, connectionInfo, createRegisterSale, retryCounter);
 };
 
 var updateConsignmentProduct = function(args, connectionInfo, retryCounter) {
   log.debug('inside updateConsignmentProduct()', args);
 
-  if ( !(args && argsAreValid(args)) ) {
+  if ( !(args && utils.argsAreValid(args)) ) {
     return Promise.reject('missing required arguments for updateConsignmentProduct()');
   }
 
@@ -2519,13 +1767,13 @@ var updateConsignmentProduct = function(args, connectionInfo, retryCounter) {
   log.debug(options.method, options.url);
   log.debug('body:', options.json);
 
-  return sendRequest(options, args, connectionInfo, updateConsignmentProduct, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, updateConsignmentProduct, retryCounter);
 };
 
 var markStockOrderAsSent = function(args, connectionInfo, retryCounter) {
   log.debug('inside markStockOrderAsSent()', args);
 
-  if ( !(args && argsAreValid(args)) ) {
+  if ( !(args && utils.argsAreValid(args)) ) {
     return Promise.reject('missing required arguments for markStockOrderAsSent()');
   }
 
@@ -2555,13 +1803,13 @@ var markStockOrderAsSent = function(args, connectionInfo, retryCounter) {
   log.debug(options.method, options.url);
   log.debug('body:', options.json);
 
-  return sendRequest(options, args, connectionInfo, markStockOrderAsSent, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, markStockOrderAsSent, retryCounter);
 };
 
 var markStockOrderAsReceived = function(args, connectionInfo, retryCounter) {
   log.debug('inside markStockOrderAsReceived()', args);
 
-  if ( !(args && argsAreValid(args)) ) {
+  if ( !(args && utils.argsAreValid(args)) ) {
     return Promise.reject('missing required arguments for markStockOrderAsReceived()');
   }
 
@@ -2591,13 +1839,13 @@ var markStockOrderAsReceived = function(args, connectionInfo, retryCounter) {
   log.debug(options.method, options.url);
   log.debug('body:', options.json);
 
-  return sendRequest(options, args, connectionInfo, markStockOrderAsReceived, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, markStockOrderAsReceived, retryCounter);
 };
 
 var deleteStockOrder = function(args, connectionInfo, retryCounter) {
   log.debug('inside deleteStockOrder()');
 
-  if ( !(args && argsAreValid(args)) ) {
+  if ( !(args && utils.argsAreValid(args)) ) {
     return Promise.reject('missing required arguments for deleteStockOrder()');
   }
 
@@ -2625,14 +1873,14 @@ var deleteStockOrder = function(args, connectionInfo, retryCounter) {
   };
   log.debug(options.method + ' ' + options.url);
 
-  return sendRequest(options, args, connectionInfo, deleteStockOrder, retryCounter);
+  return utils.sendRequest(options, args, connectionInfo, deleteStockOrder, retryCounter);
 };
 
 var deleteConsignmentProduct = function(args, connectionInfo, retryCounter) {
   log.debug('inside deleteConsignmentProduct()');
 
   log.debug(args);
-  if ( !(args && argsAreValid(args)) ) {
+  if ( !(args && utils.argsAreValid(args)) ) {
     return Promise.reject('missing required arguments for deleteStockOrder()');
   }
 
@@ -2659,102 +1907,7 @@ var deleteConsignmentProduct = function(args, connectionInfo, retryCounter) {
   };
   log.debug(options.method + ' ' + options.url);
 
-  return sendRequest(options, args, connectionInfo, deleteConsignmentProduct, retryCounter);
-};
-
-var getInitialAccessToken = function(tokenService, clientId, clientSecret, redirectUri, code, domainPrefix, state) {
-  // TODO: tweak winston logs to prefix method signature (like breadcrumbs) when logging?
-  log.debug('getInitialAccessToken - token_service: ' + tokenService);
-  log.debug('getInitialAccessToken - client Id: ' + clientId);
-  log.debug('getInitialAccessToken - client Secret: ' + clientSecret);
-  log.debug('getInitialAccessToken - redirect Uri: ' +  redirectUri);
-  log.debug('getInitialAccessToken - code: ' + code);
-  log.debug('getInitialAccessToken - domain_prefix: ' + domainPrefix);
-  log.debug('getInitialAccessToken - state: ' + state);
-
-  var tokenUrl = getTokenUrl(tokenService, domainPrefix);
-
-  var options = {
-    url: tokenUrl,
-    headers: {
-      'Accept': 'application/json'
-    },
-    form:{
-      'grant_type': 'authorization_code',
-      'client_id': clientId,
-      'client_secret': clientSecret,
-      'code': code,
-      'redirect_uri': redirectUri,
-      'state': state
-    }
-  };
-  return request.post(options)
-    .then(successHandler)
-    .catch(RateLimitingError, function(e) {// jshint ignore:line
-      log.error('A RateLimitingError error like "429 Too Many Requests" happened: '
-        + e.statusCode + ' ' + e.response.body + '\n'
-        + JSON.stringify(e.response.headers,null,2));
-    })
-    .catch(ClientError, function(e) {// jshint ignore:line
-      log.error('A ClientError happened: '
-          + e.statusCode + ' ' + e.response.body + '\n'
-        /*+ JSON.stringify(e.response.headers,null,2)
-         + JSON.stringify(e,null,2)*/
-      );
-      // TODO: add retry logic
-    })
-    .catch(function(e) {
-      log.error('getInitialAccessToken', 'An unexpected error occurred: ', e);
-    });
-};
-
-var refreshAccessToken = function(tokenService, clientId, clientSecret, refreshToken, domainPrefix) {
-  // TODO: tweak winston logs to prefix method signature (like breadcrumbs) when logging?
-  log.debug('refreshAccessToken - token service: ' + tokenService);
-  log.debug('refreshAccessToken - client Id: ' + clientId);
-  log.debug('refreshAccessToken - client Secret: ' + clientSecret);
-  log.debug('refreshAccessToken - refresh token: ' +  refreshToken);
-  log.debug('refreshAccessToken - domain prefix: ' + domainPrefix);
-
-  if ( !(tokenService && clientId && clientSecret && refreshToken) ) {
-    return Promise.reject('missing required arguments for refreshAccessToken()');
-  }
-
-  var tokenUrl = getTokenUrl(tokenService, domainPrefix);
-
-  var options = {
-    url: tokenUrl,
-    headers: {
-      'Accept': 'application/json'
-    },
-    form:{
-      'grant_type': 'refresh_token',
-      'client_id': clientId,
-      'client_secret': clientSecret,
-      'refresh_token': refreshToken
-    }
-  };
-  return request.post(options)
-    .then(successHandler)
-    .catch(RateLimitingError, function(e) {// jshint ignore:line
-      log.error('A RateLimitingError error like "429 Too Many Requests" happened: '
-        + e.statusCode + ' ' + e.response.body + '\n'
-        + JSON.stringify(e.response.headers,null,2));
-    })
-    .catch(ClientError, function(e) {// jshint ignore:line
-      log.error('A ClientError happened: '
-          + e.statusCode + ' ' + e.response.body + '\n'
-        /*+ JSON.stringify(e.response.headers,null,2)
-         + JSON.stringify(e,null,2)*/
-      );
-      // NOTE: why not throw or retry?
-      // sample: "error_description": "The refresh token is invalid."
-      // in such a case retrying just doesn't make sense, its better to fail fast
-      // which will happen when the methods calling this function try to access its results
-    })
-    .catch(function(e) {
-      log.error('refreshAccessToken', 'An unexpected error occurred: ', e);
-    });
+  return utils.sendRequest(options, args, connectionInfo, deleteConsignmentProduct, retryCounter);
 };
 
 /**
@@ -2802,7 +1955,7 @@ module.exports = function(dependencies) {
   log = dependencies.winston || require('winston');
   if (!dependencies.winston) { // if winston is being instantiated within this method then take these actions
     log.remove(log.transports.Console);
-    if (process.env.NODE_ENV !== 'test') {
+    if (process.env.NODE_ENV !== 'test') { // TODO: we want the ability to turn off these logs in production/staging too
       log.add(log.transports.Console, {
         colorize: true,
         timestamp: false,
@@ -2819,24 +1972,26 @@ module.exports = function(dependencies) {
     }
   }
 
+  // (1.5) add missing dependencies that had to be initialized
+  if (!dependencies.underscore) {dependencies.underscore = _;}
+  if (!dependencies.moment) {dependencies.moment = moment;}
+  if (!dependencies.bluebird) {dependencies.bluebird = Promise;}
+  if (!dependencies['request-promise']) {dependencies['request-promise'] = request;}
+  if (!dependencies.winston) {dependencies.winston = log;}
+
   // (2) initialize any module-scoped variables which need the dependencies
-  // ...
+  utils = require('./lib/utils.js')(dependencies);
+  dependencies.utils = utils;
+  //product = require('./lib/product.js')(dependencies);
+  product = require('./lib/product.js')(dependencies);
+  //console.log('product', product);
+  argsForInput.products = product.args;
+  //console.log('argsForInput', argsForInput);
 
   // (3) expose the SDK
   return {
     args: argsForInput,
-    products: {
-      fetch: fetchProducts,
-      fetchById: fetchProduct,
-      fetchByHandle: fetchProductByHandle,
-      fetchBySku: fetchProductBySku,
-      fetchAll: fetchAllProducts,
-      fetchPaginationInfo: fetchPaginationInfo,
-      update: updateProductById,
-      delete: deleteProductById,
-      create: createProduct,
-      uploadImage: uploadProductImage
-    },
+    products: product.endpoints,
     registers: {
       fetch: fetchRegisters,
       fetchAll: fetchAllRegisters,
@@ -2907,8 +2062,8 @@ module.exports = function(dependencies) {
       create: createSupplier
     },
     hasAccessTokenExpired: hasAccessTokenExpired,
-    getInitialAccessToken: getInitialAccessToken,
-    refreshAccessToken: refreshAccessToken,
+    getInitialAccessToken: utils.getInitialAccessToken,
+    refreshAccessToken: utils.refreshAccessToken,
     replacer: replacer
   };
 };
